# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
}

export interface UserInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
}

export type VoteInputType = typeof VoteInputType[keyof typeof VoteInputType];


export const VoteInputType = {
  up: 'up',
  down: 'down',
} as const;

export interface VoteInput {
  type: VoteInputType;
  username: string;
}

export interface VoteStats {
  upvotes: number;
  downvotes: number;
  total: number;
  percentage: number;
}

export interface VoteResult {
  success: boolean;
  message: string;
  stats: VoteStats;
  nextVoteAt: string;
}

export interface VoteCooldown {
  error: string;
  nextVoteAt: string;
  secondsRemaining: number;
}

export interface VoteStatus {
  canVote: boolean;
  /** @nullable */
  nextVoteAt: string | null;
  secondsRemaining: number;
  /** @nullable */
  lastVoteType?: string | null;
}

export interface Comment {
  id: number;
  username: string;
  content: string;
  likes: number;
  dislikes: number;
  hasStar: boolean;
  hasHeart: boolean;
  /** @nullable */
  userReaction?: string | null;
  createdAt: string;
}

export interface CommentInput {
  username: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
}

export type CommentReactionInputType = typeof CommentReactionInputType[keyof typeof CommentReactionInputType];


export const CommentReactionInputType = {
  like: 'like',
  dislike: 'dislike',
} as const;

export interface CommentReactionInput {
  username: string;
  type: CommentReactionInputType;
}

export interface ReactionResult {
  action: string;
  /** @nullable */
  type?: string | null;
}

export interface AdminVote {
  id: number;
  ipAddress: string;
  username: string;
  voteType: string;
  createdAt: string;
}

export interface AdminAction {
  adminUsername: string;
}

export interface AdminToggleResult {
  success: boolean;
  hasStar?: boolean;
  hasHeart?: boolean;
}

export type ListCommentsParams = {
/**
 * Current user's username to include their own reaction
 */
username?: string;
};