# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Workly Game Core API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: votes
    description: Voting system
  - name: comments
    description: Comments system
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new unique username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken or reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{username}:
    get:
      operationId: getUser
      tags: [users]
      summary: Check if a username exists
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /votes:
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote for or against the project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "429":
          description: Cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteCooldown"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /votes/stats:
    get:
      operationId: getVoteStats
      tags: [votes]
      summary: Get vote statistics
      responses:
        "200":
          description: Vote statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteStats"

  /votes/status:
    get:
      operationId: getVoteStatus
      tags: [votes]
      summary: Check if current IP can vote
      responses:
        "200":
          description: Vote status for current IP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteStatus"

  /comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List all comments with reaction counts
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
          description: Current user's username to include their own reaction
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Post a new comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /comments/{id}/react:
    post:
      operationId: reactToComment
      tags: [comments]
      summary: Like or dislike a comment (toggles off if same type)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentReactionInput"
      responses:
        "200":
          description: Reaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/votes:
    get:
      operationId: adminListVotes
      tags: [admin]
      summary: List all votes with user and IP info (admin only)
      parameters:
        - name: x-admin-username
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All votes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminVote"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/comments/{id}/star:
    post:
      operationId: adminToggleStar
      tags: [admin]
      summary: Toggle star on a comment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAction"
      responses:
        "200":
          description: Star toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToggleResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/comments/{id}/heart:
    post:
      operationId: adminToggleHeart
      tags: [admin]
      summary: Toggle heart on a comment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAction"
      responses:
        "200":
          description: Heart toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToggleResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/comments/{id}:
    delete:
      operationId: adminDeleteComment
      tags: [admin]
      summary: Delete a comment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAction"
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, createdAt]

    UserInput:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
      required: [username]

    VoteInput:
      type: object
      properties:
        type:
          type: string
          enum: [up, down]
        username:
          type: string
      required: [type, username]

    VoteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        stats:
          $ref: "#/components/schemas/VoteStats"
        nextVoteAt:
          type: string
          format: date-time
      required: [success, message, stats, nextVoteAt]

    VoteCooldown:
      type: object
      properties:
        error:
          type: string
        nextVoteAt:
          type: string
          format: date-time
        secondsRemaining:
          type: integer
      required: [error, nextVoteAt, secondsRemaining]

    VoteStats:
      type: object
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
        total:
          type: integer
        percentage:
          type: number
      required: [upvotes, downvotes, total, percentage]

    VoteStatus:
      type: object
      properties:
        canVote:
          type: boolean
        nextVoteAt:
          type: ["string", "null"]
          format: date-time
        secondsRemaining:
          type: integer
        lastVoteType:
          type: ["string", "null"]
      required: [canVote, nextVoteAt, secondsRemaining]

    Comment:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        content:
          type: string
        likes:
          type: integer
        dislikes:
          type: integer
        hasStar:
          type: boolean
        hasHeart:
          type: boolean
        userReaction:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, username, content, likes, dislikes, hasStar, hasHeart, createdAt]

    CommentInput:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
          minLength: 1
          maxLength: 500
      required: [username, content]

    CommentReactionInput:
      type: object
      properties:
        username:
          type: string
        type:
          type: string
          enum: [like, dislike]
      required: [username, type]

    ReactionResult:
      type: object
      properties:
        action:
          type: string
        type:
          type: ["string", "null"]
      required: [action]

    AdminVote:
      type: object
      properties:
        id:
          type: integer
        ipAddress:
          type: string
        username:
          type: string
        voteType:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ipAddress, username, voteType, createdAt]

    AdminAction:
      type: object
      properties:
        adminUsername:
          type: string
      required: [adminUsername]

    AdminToggleResult:
      type: object
      properties:
        success:
          type: boolean
        hasStar:
          type: boolean
        hasHeart:
          type: boolean
      required: [success]