# lib/api-zod/src/generated/types/adminVote.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface AdminVote {
  id: number;
  ipAddress: string;
  username: string;
  voteType: string;
  createdAt: Date;
}