# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface Comment {
  id: number;
  username: string;
  content: string;
  likes: number;
  dislikes: number;
  hasStar: boolean;
  hasHeart: boolean;
  /** @nullable */
  userReaction?: string | null;
  createdAt: Date;
}