# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface CommentInput {
  username: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
}