# lib/api-zod/src/generated/types/commentReactionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */
import type { CommentReactionInputType } from './commentReactionInputType';

export interface CommentReactionInput {
  username: string;
  type: CommentReactionInputType;
}