# lib/api-zod/src/generated/types/reactionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface ReactionResult {
  action: string;
  /** @nullable */
  type?: string | null;
}