# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: number;
  username: string;
  createdAt: Date;
}