# lib/api-zod/src/generated/types/voteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */
import type { VoteInputType } from './voteInputType';

export interface VoteInput {
  type: VoteInputType;
  username: string;
}