# lib/api-zod/src/generated/types/voteStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Workly Game Core API
 * OpenAPI spec version: 0.2.0
 */

export interface VoteStatus {
  canVote: boolean;
  /** @nullable */
  nextVoteAt: Date | null;
  secondsRemaining: number;
  /** @nullable */
  lastVoteType?: string | null;
}